+++ src/main/java/com/univ/backend/controllers/TeamController.java
@@ -1,12 +1,12 @@
 package com.univ.backend.controllers;
 
+import com.univ.backend.entities.Admin;
 import com.univ.backend.entities.TeamEntity;
-import com.univ.backend.exceptions.ImageFormatException;
-import com.univ.backend.exceptions.MandatoryFieldFoundEmptyException;
-import com.univ.backend.exceptions.TeamMemberNotFoundException;
+import com.univ.backend.exceptions.*;
 import com.univ.backend.response.TeamDeleteRequestResponse;
 import com.univ.backend.response.TeamMemberUpdateResponse;
 import com.univ.backend.response.TeamPutRequestResponse;
+import com.univ.backend.services.AdminService;
 import com.univ.backend.services.TeamService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -19,7 +19,6 @@ import java.util.Calendar;
 import java.util.List;
 
 
-
 @RestController
 @RequestMapping("/api/v1")
 //@CrossOrigin(origins = "https://localhost:5173")
@@ -28,17 +27,33 @@ public class TeamController {
     @Autowired
     public TeamService service;
 
+    @Autowired
+    private AdminService adminService;
+
     @PostMapping("/team")
-    public TeamPutRequestResponse postTeamDetails(@RequestParam("image") MultipartFile image, @RequestParam("name") String name, @RequestParam("role") String role, @RequestParam("info") String info, @RequestParam("url") String url) throws MandatoryFieldFoundEmptyException, IOException, ImageFormatException {
-        return service.addTeam(
-                TeamEntity.builder()
-                        .name(name)
-                        .role(role)
-                        .info(info)
-                        .url(url)
-                        .build(),
-                image
-        );
+    public TeamPutRequestResponse postTeamDetails(
+            @RequestParam("image") MultipartFile image,
+            @RequestParam("name") String name,
+            @RequestParam("role") String role,
+            @RequestParam("info") String info,
+            @RequestParam("url") String url,
+            @RequestParam("username") String adminUserName,
+            @RequestParam("password") String adminPassword
+    ) throws MandatoryFieldFoundEmptyException, IOException, ImageFormatException, AdminNotFoundException, IncorrectAdminDataException {
+        assert (adminUserName != null && adminPassword != null) : "Admin username and password cannot be null.";
+        if (adminService.verifyAdmin(adminUserName, adminPassword)) {
+            return service.addTeam(
+                    TeamEntity.builder()
+                            .name(name)
+                            .role(role)
+                            .info(info)
+                            .url(url)
+                            .build(),
+                    image
+            );
+        } else {
+            throw new IncorrectAdminDataException("The provided admin data was incorrect!", new Admin(adminUserName, adminPassword));
+        }
     }
 
     @GetMapping("/team")
@@ -56,28 +71,44 @@ public class TeamController {
             @RequestParam(value = "name", required = false) String name,
             @RequestParam(value = "role", required = false) String role,
             @RequestParam(value = "info", required = false) String info,
-            @RequestParam(value = "url", required = false) String url
-    ) throws TeamMemberNotFoundException, IOException, ImageFormatException {
-        TeamEntity teamMember = service.updateTeamDetails(
-                TeamEntity.builder()
-                        .name(name)
-                        .role(role)
-                        .info(info)
-                        .url(url)
-                        .build(),
-                fetchName,
-                image);
-        return new TeamMemberUpdateResponse(
-                HttpStatus.OK,
-                teamMember,
-                teamMember.getId() + " was updated successfully!",
-                Calendar.getInstance().getTime().getTime());
+            @RequestParam(value = "url", required = false) String url,
+            @RequestParam("username") String adminUserName,
+            @RequestParam("password") String adminPassword
+    ) throws TeamMemberNotFoundException, IOException, ImageFormatException, AdminNotFoundException, IncorrectAdminDataException {
+        assert (adminUserName != null && adminPassword != null) : "Admin username and password cannot be null.";
+        if (adminService.verifyAdmin(adminUserName, adminPassword)) {
+            TeamEntity teamMember = service.updateTeamDetails(
+                    TeamEntity.builder()
+                            .name(name)
+                            .role(role)
+                            .info(info)
+                            .url(url)
+                            .build(),
+                    fetchName,
+                    image);
+            return new TeamMemberUpdateResponse(
+                    HttpStatus.OK,
+                    teamMember,
+                    teamMember.getId() + " was updated successfully!",
+                    Calendar.getInstance().getTime().getTime());
+        } else {
+            throw new IncorrectAdminDataException("The provided admin data was incorrect!", new Admin(adminUserName, adminPassword));
+        }
     }
 
     @DeleteMapping("/team/{name}")
-    public TeamDeleteRequestResponse teamMemberDeleteRequest(@PathVariable(name = "name") String name) throws TeamMemberNotFoundException, FileNotFoundException {
-        TeamEntity deletedEntity = service.deleteTeamMemberById(name);
-        return new TeamDeleteRequestResponse(HttpStatus.OK, deletedEntity, "Member " + name + " deleted successfully!", Calendar.getInstance().getTime().getTime());
+    public TeamDeleteRequestResponse teamMemberDeleteRequest(
+            @PathVariable(name = "name") String name,
+            @RequestParam("username") String adminUserName,
+            @RequestParam("password") String adminPassword
+    ) throws TeamMemberNotFoundException, FileNotFoundException, IncorrectAdminDataException, AdminNotFoundException {
+        assert (adminUserName != null && adminPassword != null) : "Admin username and password cannot be null.";
+        if(adminService.verifyAdmin(adminUserName, adminPassword)) {
+            TeamEntity deletedEntity = service.deleteTeamMemberById(name);
+            return new TeamDeleteRequestResponse(HttpStatus.OK, deletedEntity, "Member " + name + " deleted successfully!", Calendar.getInstance().getTime().getTime());
+        } else {
+            throw new IncorrectAdminDataException("The provided admin data was incorrect!", new Admin(adminUserName, adminPassword));
+        }
     }
 
 }
